docs: Use sections in hello world tutorial (#132)
Broke up the hello world tutorial into steps, each with a small heading,
a description of the task and then the code to use. This makes it
easier to follow along and also makes it easier to find the relevant
code when looking for a specific part of the tutorial.

Also tweaked the information about the imports to make it clearer why
they are needed, and added information about how the Stylize shorthands
are used. Added more direct links to documentation to make it easy to
find more information about the various parts of the code.

Also added a link directly to the code in the repo to make it easy for
people to view the docs and code in a side-by-side view.