Revampt the `components` template's documentation. (#941)
I updated the `main.rs` and `project-structure` to be up to date. I'll
proceed with all new docs in a filename-new file. When everything is
finalised, I'll update the main pages accordingly.

Please give feedback on the new documentation, what is missing, and what
needs to be removed. Thank you!

---------

Co-authored-by: Orhun Parmaksız <orhunparmaksiz@gmail.com>